+++ pmd-html/src/test/java/net/sourceforge/pmd/lang/html/HtmlXPathRuleTest.java
@@ -14,7 +14,7 @@ import net.sourceforge.pmd.lang.html.ast.ASTHtmlComment;
 import net.sourceforge.pmd.lang.html.ast.ASTHtmlDocument;
 import net.sourceforge.pmd.lang.html.ast.ASTHtmlTextNode;
 import net.sourceforge.pmd.lang.html.ast.HtmlParsingHelper;
-import net.sourceforge.pmd.lang.rule.XPathRule;
+import net.sourceforge.pmd.lang.rule.xpath.XPathRule;
 import net.sourceforge.pmd.lang.rule.xpath.XPathVersion;
 import net.sourceforge.pmd.reporting.Report;
 import net.sourceforge.pmd.reporting.RuleViolation;